Add files via upload
According to the notes provided, we first import rospy and the standard Int16/Int32 message type from std_msgs.msg.
We then start by writing a publisher function, where we publish multiples of 2 on a topic called "multiple_of_2" using a while loop.
We use rospy.Rate(n) to control how fast the numbers are published 
We also use rospy.loginfo() to print the value being published.
Then, we create a first subscriber node. It listens to the "multiple_of_2" topic using a callback() function. In this function, we multiply the received number by 10 and publish it to a new topic called "multiplied_by_10".
We use rospy.sleep(n) so that there is a time lag of n sec
After that, we build a second subscriber node, which listens to the "multiplied_by_10" topic. Inside its callback, we take the number received, add 5 to it, and print the final result using rospy.loginfo().
The if_name == '_main_' thing is the default as provided in the notes.