Add files via upload
According to the notes, we start by importing rospy, and then we import our message from the package .
We first create a subscriber node (the listener) where we define a callback() function. This function is called every time a new bot position message is received. Inside the callback, we use rospy.loginfo to display the bot’s position in the format (x, y, direction).
Then, we create a publisher node where the user enters movement commands like forward, left, right.......
We maintain a track of the bot’s current x, y position and its direction. Based on the command entered, we update these variables accordingly.
After updating the position, we publish a bot_position message with the new x, y, and direction.
We use rospy.sleep(n) so that there is a time lag of n sec
The if_name == '_main_' thing is the default as provided in the notes.